Update layout to be more portable across views
